Fixed problem with nonfree python test: could not find local test images
